New / Choices / Errors
Through alternative methods, this is my progress so far.

To my knowledge you cannot skip scenes or it'll fail to load the mod properly. This change allows for the button to show properly and is supposed to loop around to show you a new button called "Take a walk." This will allow you to be able to run into Kevin and meet up with him via Chapter 1. From there you can do more by being able to spend time with him.

ERRORS:
- "Take a walk." Choice option does not load the scene with Kevin rather stays on the menu not allowing you to encounter Kevin when the Choice has been selected for usage.