Official Relase of SparkzTehDragon's Modpack Start | Beta
This is the first modpack for AWSW that I have created with friends and acquaintences!

If there is a problem my Discord is in the README.txt file within this commit. This is in Version 0.1 and will be updated as I progress in adding it to the game. There are ways to trouble shoot it in a way that will allow it to work. Be prepared! Be safe out there! This is a rather short scene currently but it is basically a duplicate from the xkevin.rpy file in the game itself and it will and should be extended further more in a few more days that will allow you to play it more without difficulty or boredom.