New Files / New Update - SEBASTIAN'S UPDATE - BETA UPDATE
So, for starters Sebastian has received an update that will allow the Main Character to meet with him in the 1st chapter. To be able to see his WIP Scene you'll need to go to the second "Meet with Sebastian" it should be the one closest to the bottom of the list.

Also to note: When the game itself throws a LOCAL VARIABLE error such as "NOT DEFINED" just press "IGNORE" it should go rollover it without any issue.

Stay tuned for more! Sorry for the late update a month from now, I'll still be working on the mod in the meantime while continuing doing my commissions. Thanks for taking the time to download the mod itself for BETA.